\name{vb_model2_la}
\alias{vb_model2_la}

\title{Variational Bayes (VB) Laplace approximation to the Royle Dorazio (2008) model}

\description{
The following function undertakes a VB Laplace approximation to the Royle Dorazio (2008) model. We assume that all link functions are logistic.}

\usage{
vb_model2_la(formula, design_mats, alpha_0, beta_0, Sigma_alpha_0, Sigma_beta_0, LargeSample = TRUE, epsilon = 1e-05)
}

\arguments{
  \item{}{}
  \item{formula }{ - Double right-hand side formula describing covariates of detection and occupancy in that order.}
  \item{design_mats }{ - A named list generated by the call vb_Designs(W, X, y). 
  
  W is a named list of data.frames of covariates that vary within sites. i.e. The dataframes are of dimension n by J where each row is associated with a site and each column represents a site visit. 
  
  e.g. Suppose W contained three dataframes W1, W2 and W3; W$W1[1, ] = the covariate values for site 1 for all of the visits. Note that some of the entries might be 'NA' meaning that no visit took place at those occasions. 

X is a named data.frame that varies at site level.

y is an n by J matrix of the detection, non-detection data, where n is the number of sites, J is the maximum number of sampling periods per site.

NOTE: THE FUNCTION DOES NOT ALLOW THERE TO BE ANY MISSING VALUES IN THE COVARIATE MATRICES IF A SURVEY WAS UNDERTAKEN AT A PARTICULAR LOCATION!!!}
  \item{alpha_0 }{ - Prior mean of the detection covariate coefficients. It is assumed that the prior distribution of  \eqn{ \alpha ~N(alpha_0, Sigma_alpha_0)}. Here \eqn{\alpha} is viewed as a vector.}
  \item{beta_0 }{ - Prior mean of the occurence covariate coefficients. It is assumed that the prior distribution of \eqn{ \beta~ N(beta_0, Sigma_beta_0)}. Here \eqn{\beta} is viewed as a vector.}
  \item{Sigma_alpha_0 }{ - Prior covariance matrix of the detection covariate coefficients.}
  \item{Sigma_beta_0 }{ - Prior covariance matrix of the occurence covariate coefficients. }
  \item{LargeSample }{ - LargeSample==TRUE - indicates that the number of sites is 'large' and that an approximation to \eqn{B(\mu, \sigma^2)} is used instead of integrations. (otherwise numerical integrations are performed)}
  \item{epsilon }{ - Convergence measured relative to this quantity.}
}

\details{}

\value{
  \item{}{}
  \item{alpha }{The VB estimate of the posterior mean vector of \eqn{\alpha}. (p by 1 vector)}
  \item{beta }{The VB estimate of the posterior mean vector of \eqn{\beta}. (q by 1 vector)}
  \item{Sigma_alpha }{The VB estimate of the posterior covariance matrix of the \eqn{\alpha} vector. (p by p matrix)}
  \item{Sigma_beta }{The VB estimate of the posterior covariance matrix of the \eqn{\beta} vector. (q by q matrix)}
  \item{occup_p }{The VB estimate of the posterior occupancy probabilities at the sites considered. (n by 1 vector)}
  \item{Log_mla }{The lower bound of the log marginal loglikelihood.}  
}

\references{
\itemize{
\item Dorazio, R. M. and Rodriguez, D. T. (2012), A Gibbs sampler for Bayesian analysis of site-occupancy data. Methods in Ecology and Evolution, 3: 1093-1098. \url{doi: 10.1111/j.2041-210X.2012.00237.x}
\item MacKenzie, D.I., Nichols, J.D., Lachman, G.B., Droege, S., Royle, J.A., Langtimm, C.A. (2002). Estimating site occupancy rates when detection probabilities are less than one. Ecology, 83(8): 2248-2255
\item Royle, J. A., & Dorazio, R. M. (2008). Hierarchical modeling and inference in ecology: the analysis of data from populations, metapopulations and communities. Academic Press.
}}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}

\examples{
## Load the package as well as the data
##-------------------------------------
require(vboccupancy)
data(SimData2)

#Set Uninformative priors
#------------------------
#Coefficients in the detection model
alpha_0 <- matrix(0, ncol=1, nrow=4)
#Covariance matrix of the coefficients in the detection model
Sigma_alpha_0 <- diag(4)*1000
#Coefficients in the occupancy process
beta_0 <- matrix(0, ncol=1, nrow=3)
#Covariance matrix of the coefficients in the occupancy model
Sigma_beta_0 <- diag(3)*1000

#Construct the required matrices using vb_Designs
#------------------------------------------------
design_mats<-vb_Designs(W=SimData2$W.eg.l1, X=SimData2$X, y=SimData2$y)

#Here we use the large sample approximation and run the VB algorithm
#-------------------------------------------------------------------
#Assume that the formula used will be of the following form: formula1<- y~X1+X2~W1+W2+W3
#The occupancy model uses 2 covariates, X1 and X2; while
#the detection model uses 3 covariates W1, W2 and W3
#Intercepts are included in both models
#The function does not allow one to repress the intercept term

vb_fit<-vb_model2_la(y~X1+X2~W1+W2+W3, design_mats=design_mats, 
                     alpha_0=alpha_0, beta_0=beta_0, 
                     Sigma_alpha_0=Sigma_alpha_0, Sigma_beta_0=Sigma_beta_0,
                     LargeSample=TRUE, epsilon=1e-5)

#The detection model parameters
vb_fit$alpha

#The occupancy model parameters
vb_fit$beta

#The respective covariance matrices
vb_fit$Sigma_alpha
vb_fit$Sigma_beta

#The approximate conditional ocupancy probabilities
plot(vb_fit$occup_p)

}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
