\name{vboccupancy-package}
\alias{vboccupancy-package}
\alias{vboccupancy}
\docType{package}

\title{vboccupancy}

\description{The package provides code in order to fit a Variational Bayes approximation to the occupancy models developed by MacKenzie et al (2002).}

\details{
\tabular{ll}{
Package: \tab vboccupancy\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-07\cr
License: \tab GPL-2\cr
}
%An overview of how to use the package, including the most important functions 
}

\author{A. E. Clark

Maintainer: <allan dot clark at uct dot ac dot za> 
}
\references{
\itemize{
%%\item Dorazio, R. M. and Rodriguez, D. T. (2012), A Gibbs sampler for Bayesian analysis of site-occupancy data. Methods in Ecology and Evolution, 3: 1093-1098. \url{doi: 10.1111/j.2041-210X.2012.00237.x}
\item MacKenzie, D.I., Nichols, J.D., Lachman, G.B., Droege, S., Royle, J.A., Langtimm, C.A. (2002). Estimating site occupancy rates when detection probabilities are less than one. Ecology, 83(8): 2248-2255
\item Royle, J. A., & Dorazio, R. M. (2008). Hierarchical modeling and inference in ecology: the analysis of data from populations, metapopulations and communities. Academic Press.}
}

\keyword{ package }


